+++ features/step_definitions/element_explicit_waiting_steps.rb
@@ -13,8 +13,7 @@ end
 Then("an exception is raised when I wait for an element that won't appear") do
   start_time = Time.now
 
-  expect { @test_site.slow.last_link(wait: lower_bound_delay) }
-    .to raise_error(Capybara::ElementNotFound)
+  expect { @test_site.slow.last_link(wait: lower_bound_delay) }.to raise_error(Capybara::ElementNotFound)
 
   expect(Time.now - start_time).to be < time_delay
 end
@@ -25,13 +24,11 @@ Then('I get an error when I wait for an element to vanish within the limit') do
 end
 
 Then("an exception is raised when I wait for an element that won't vanish") do
-  expect { @test_site.home.wait_until_header_invisible }
-    .to raise_error(SitePrism::ElementInvisibilityTimeoutError)
+  expect { @test_site.home.wait_until_header_invisible }.to raise_error(SitePrism::ElementInvisibilityTimeoutError)
 end
 
 Then('I can wait a variable time for elements to disappear') do
-  expect { @test_site.vanishing.removed_elements(wait: upper_bound_delay, count: 0) }
-    .not_to raise_error
+  expect { @test_site.vanishing.removed_elements(wait: upper_bound_delay, count: 0) }.not_to raise_error
 
   expect(@test_site.vanishing).to have_no_removed_elements
 end
@@ -46,8 +43,7 @@ Then('I get a timeout error when waiting for an element to become visible within
 end
 
 Then('I get a timeout error when waiting for an element with default limit') do
-  expect { @test_site.slow.wait_until_invisible_visible }
-    .to raise_error(SitePrism::ElementVisibilityTimeoutError)
+  expect { @test_site.slow.wait_until_invisible_visible }.to raise_error(SitePrism::ElementVisibilityTimeoutError)
 end
 
 When('I wait until a particular element is visible') do
@@ -85,8 +81,7 @@ end
 Then('I can override the wait time using a Capybara.using_wait_time block') do
   start_time = Time.now
   Capybara.using_wait_time(lower_bound_delay) do
-    expect { @test_site.slow.last_link }
-      .to raise_error(Capybara::ElementNotFound)
+    expect { @test_site.slow.last_link }.to raise_error(Capybara::ElementNotFound)
   end
 
   expect(Time.now - start_time).to be_between(lower_bound_delay, upper_bound_delay)
@@ -100,7 +95,5 @@ Then('I am not made to wait to check a nonexistent element for invisibility') do
 end
 
 Then('an error is thrown when waiting for an element in a vanishing section') do
-  expect do
-    @test_site.vanishing.container.wait_until_embedded_invisible
-  end.to raise_error(Capybara::ElementNotFound)
+  expect { @test_site.vanishing.container.wait_until_embedded_invisible }.to raise_error(Capybara::ElementNotFound)
 end