+++ spec/site_prism/dsl/validator_spec.rb
@@ -1,9 +1,9 @@
 # frozen_string_literal: true
 
-describe SitePrism::DSLValidator do
+describe SitePrism::DSL::Validator do
   let(:validator) do
     Class.new do
-      extend SitePrism::DSLValidator
+      extend SitePrism::DSL::Validator
     end
   end
 
@@ -41,7 +41,7 @@ describe SitePrism::DSLValidator do
     context 'with a valid string' do
       subject { validator.invalid?('abcdef123_XYZ') }
 
-      it { is_expected.to be false }
+      it { is_expected.to be_nil }
     end
   end
 end