+++ spec/site_prism/iframe_spec.rb
@@ -58,6 +58,7 @@ describe 'Iframe' do
       end
     end
 
+    # Stop the $stdout process leaking cross-tests
     before { wipe_logger! }
 
     it 'will throw a warning when creating an iFrame with an ambiguous locator' do
@@ -66,7 +67,7 @@ describe 'Iframe' do
         invalid_class
       end
 
-      expect(lines(log_messages)).to eq 4
+      expect(lines(log_messages)).to be_positive
     end
   end